Split out generating serialisation and deserialisation functions
